refactor: remove grid demo module
