+++ vaadin-grid-flow-parent/vaadin-grid-flow-integration-tests/src/main/java/com/vaadin/flow/component/grid/it/data/CountryData.java
@@ -1,4 +1,4 @@
-package com.vaadin.flow.component.grid.demo.data;
+package com.vaadin.flow.component.grid.it.data;
 
 import java.util.ArrayList;
 import java.util.List;
